* origin/released:
  libX11 libxcb mesa xserver xcb-proto xkeyboard-config git update 4 Mar 2014

Conflicts:
	mesalib/src/mapi/glapi/glapi.h
	mesalib/src/mapi/glapi/glthread.c
	mesalib/src/mesa/drivers/dri/common/dri_util.c
	mesalib/src/mesa/main/bufferobj.c
	xorg-server/dix/dispatch.c
	xorg-server/hw/xwin/glx/gen_gl_wrappers.py
	xorg-server/hw/xwin/winmultiwindowwm.c

+++ xorg-server/hw/xfree86/os-support/linux/lnx_video.c
@@ -204,7 +204,7 @@ mtrr_cull_wc_region(int screenNum, unsigned long base, unsigned long size,
             xf86DrvMsgVerb(screenNum, X_WARNING, 0,
                            "Failed to remove MMIO "
                            "write-combining range (0x%lx,0x%lx)\n",
-                           gent.base, (unsigned long) gent.size);
+                           (unsigned long)gent.base, (unsigned long) gent.size);
         }
     }
     return wcreturn;
